requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence_transformers
langchain_chroma
langchainhub
unstructured